The current state of the project

+++ docker/prepare_context.sh
@@ -9,16 +9,20 @@ set -euo pipefail
 refinery_version="$(./get_version.sh)"
 cli_distribution_name="refinery-generator-cli-${refinery_version}"
 web_distribution_name="refinery-language-web-${refinery_version}"
+generator_distribution_name="refinery-generator-server-${refinery_version}"
 
-rm -rf "${cli_distribution_name}" "${web_distribution_name}" {cli,web}_dist context/extracted
+rm -rf "${cli_distribution_name}" "${web_distribution_name}" \
+ "${generator_distribution_name}" {cli,web,generator}_dist context/extracted
 
 tar -xvf "../subprojects/generator-cli/build/distributions/${cli_distribution_name}.tar"
 mv "${cli_distribution_name}" cli_dist
 tar -xvf "../subprojects/language-web/build/distributions/${web_distribution_name}.tar"
 mv "${web_distribution_name}" web_dist
-mkdir -p context/extracted/{cli,web}_{,app_}lib \
+tar -xvf "../subprojects/generator-server/build/distributions/${generator_distribution_name}.tar"
+mv "${generator_distribution_name}" generator_dist
+mkdir -p context/extracted/{cli,web,generator}_{,app_}lib \
     context/extracted/common_{,amd64_,arm64_}lib \
-    context/extracted/{cli,web}_{amd64,arm64}_bin
+    context/extracted/{cli,web,generator}_{amd64,arm64}_bin
 
 move_application_jars() {
     prefix="$1"
@@ -29,12 +33,20 @@ move_application_jars() {
 
 move_application_jars cli
 move_application_jars web
+move_application_jars generator
 
+# Dependency deduplication for optimizing Docker image creation
+# It used to be checking whether a jar file is in atleast two locations (cli/web/generator)
+  # However, that caused almost all files to be moved to the base image
+# It checks whether a file from cli is in all 3 of the target libs
+# TODO modify in the future. Maybe refactor the generator codebase, so that it isn't as tightly coupled
 for i in cli_dist/lib/*; do
     j="web${i#cli}"
-    if [[ -f "$j" ]]; then
+    generator_file="generator${i#cli}"
+    if [[ -f "$j" && -f "$generator_file" ]]; then
         mv "$i" "context/extracted/common_lib${i#cli_dist/lib}"
         rm "$j"
+	rm "$generator_file"
     fi
 done
 
@@ -62,5 +74,6 @@ prepare_application() {
 
 prepare_application cli generator-cli
 prepare_application web language-web
+prepare_application generator generator-server
 
-rm -rf {cli,web}_dist
+rm -rf {cli,web,generator}_dist
